Fix all ESLint warnings and improve code quality
- Replace TypeScript 'any' types with proper type definitions
- Fix unused variables in monitoring.ts
- Resolve React refresh warnings by moving constants to separate files
- Update ESLint configuration with React version detection
- All linting warnings resolved, build passes successfully
