+++ app/components/PerformanceOptimizer.tsx
@@ -54,11 +54,11 @@ export const PerformanceOptimizer: React.FC<PerformanceOptimizerProps> = ({ chil
               console.log('LCP:', entry.startTime);
             }
             if (entry.entryType === 'first-input') {
-              const fidEntry = entry as any;
+              const fidEntry = entry as PerformanceEventTiming;
               console.log('FID:', fidEntry.processingStart - fidEntry.startTime);
             }
             if (entry.entryType === 'layout-shift') {
-              console.log('CLS:', (entry as any).value);
+              console.log('CLS:', (entry as PerformanceEntry & { value: number }).value);
             }
           });
         });
