+++ app/constants/viewport.ts
@@ -1,7 +1,6 @@
-import type { Viewport } from 'next'
-
-export const viewport: Viewport = {
+export const viewport = {
   width: 'device-width',
   initialScale: 1,
   maximumScale: 1,
+  userScalable: false,
 };
