+++ app/hooks/usePerformanceMetrics.ts
@@ -44,7 +44,7 @@ export const usePerformanceMetrics = () => {
     const fidObserver = new PerformanceObserver((list) => {
       const entries = list.getEntries();
       entries.forEach((entry: PerformanceEntry) => {
-        const fidEntry = entry as any;
+        const fidEntry = entry as PerformanceEventTiming;
         setMetrics(prev => ({ ...prev, fid: fidEntry.processingStart - fidEntry.startTime }));
       });
     });
@@ -55,7 +55,7 @@ export const usePerformanceMetrics = () => {
     const clsObserver = new PerformanceObserver((list) => {
       const entries = list.getEntries();
       entries.forEach((entry: PerformanceEntry) => {
-        const clsEntry = entry as any;
+        const clsEntry = entry as PerformanceEntry & { hadRecentInput?: boolean; value: number };
         if (!clsEntry.hadRecentInput) {
           clsValue += clsEntry.value;
         }
